Separate ContactTestResultCallback
